fix(misc): rename pkgs
